requirements.txt
langchain
langchain-openai
langchain-core
langgraph>0.2.27
python-dotenv
huggingface-hub
transformers
requests
python-dotenv
ipykernel
together
bs4
langchain-text-splitters
langchain-community
unstructured
libmagic
ftfy
numpy
tiktoken
sentence_transformers
torch
tqdm
ollama
faiss-cpu
langchain-huggingface
ollama
streamlit
langchain_ollama
